+++ contrib/opencensus-ext-django/tests/test_django_middleware.py
@@ -23,9 +23,9 @@ from opencensus.ext.jaeger import trace_exporter as jaeger_exporter
 from opencensus.ext.ocagent import trace_exporter as ocagent_exporter
 from opencensus.ext.zipkin import trace_exporter as zipkin_exporter
 from opencensus.trace import execution_context
+from opencensus.trace import print_exporter
 from opencensus.trace import span as span_module
 from opencensus.trace import utils
-from opencensus.trace.exporters import print_exporter
 from opencensus.trace.propagation import google_cloud_format
 from opencensus.trace.samplers import always_on
 from opencensus.trace.samplers import probability
