+++ contrib/opencensus-ext-pyramid/tests/test_pyramid_middleware.py
@@ -26,8 +26,8 @@ from opencensus.common.transports import sync
 from opencensus.ext.pyramid import pyramid_middleware
 from opencensus.ext.zipkin import trace_exporter as zipkin_exporter
 from opencensus.trace import execution_context
+from opencensus.trace import print_exporter
 from opencensus.trace import span as span_module
-from opencensus.trace.exporters import print_exporter
 from opencensus.trace.propagation import google_cloud_format
 from opencensus.trace.samplers import always_on
 from opencensus.trace.blank_span import BlankSpan
