+++ docs/trace/api/logging_exporter.rst
@@ -1,6 +1,6 @@
 Exporter - Logging Exporter
 ===========================
 
-.. automodule:: opencensus.trace.exporters.logging_exporter
+.. automodule:: opencensus.trace.logging_exporter
   :members:
   :show-inheritance:
