+++ opencensus/common/transports/async_.py
@@ -31,7 +31,7 @@ _WORKER_TERMINATOR = object()
 class _Worker(object):
     """A background thread that exports batches of data.
 
-    :type exporter: :class:`~opencensus.trace.exporters.base.Exporter` or
+    :type exporter: :class:`~opencensus.trace.base_exporter.Exporter` or
                     :class:`~opencensus.stats.exporters.base.StatsExporter`
     :param exporter: Instances of Exporter objects. Defaults to
                     :class:`.PrintExporter`. The rest options are
@@ -183,7 +183,7 @@ class _Worker(object):
 class AsyncTransport(base.Transport):
     """Asynchronous transport that uses a background thread.
 
-    :type exporter: :class:`~opencensus.trace.exporters.base.Exporter` or
+    :type exporter: :class:`~opencensus.trace.base_exporter.Exporter` or
                     :class:`~opencensus.stats.exporters.base.StatsExporter`
     :param exporter: Instances of Exporter objects. Defaults to
                      :class:`.PrintExporter`. The rest options are
