+++ tests/unit/trace/exporters/test_logging_exporter.py
@@ -17,9 +17,9 @@ import unittest
 
 import mock
 
+from opencensus.trace import logging_exporter
 from opencensus.trace import span_context
 from opencensus.trace import span_data as span_data_module
-from opencensus.trace.exporters import logging_exporter
 
 
 class TestLoggingExporter(unittest.TestCase):
